Ensure internal miner respects user input for stop/quit.
With the longer loop length, the internal miner wasn't respecting
the user's command to stop mining or exit the wallet/daemon until
the entire loop finished (or a block was received).

This change adds additional guards and checks to ensure the miner
is always aware of fGenerate and minerStopFlag, and to immediately
cease mining if either condition requires it to do so.

Amends pull #69